Conflicts:
	mih/core/src/main/java/de/mih/core/engine/tilemap/Tile.java
	mih/core/src/main/java/de/mih/core/engine/tilemap/Tilemap.java
	mih/core/src/main/java/de/mih/core/engine/tilemap/borders/BorderCollider.java
	mih/core/src/main/java/de/mih/core/game/MiH.java
	mih/core/src/main/java/de/mih/core/game/input/InGameInput.java
	mih/core/src/main/java/de/mih/core/game/systems/ControllerSystem.java
	mih/desktop/assets/maps/map1.xml

+++ mih/core/src/main/java/de/mih/core/game/components/PositionC.java
@@ -14,7 +14,7 @@ public class PositionC extends Component {
 	public final static String name = "position";
 
 	public Vector3 position;
-	public int angle;
+	public float angle;
 	
 	public PositionC()
 	{
